push for test v2

+++ src/main/java/com/example/be_java_hisp_w25_g01/dto/response/FollowersCountDTO.java
@@ -13,6 +13,6 @@ public class FollowersCountDTO {
     //US 0002
     Integer user_id;
     String user_name;
-    Integer followers_count;
+    Long followers_count;
 
 }
